replace webdav username, password with access token

+++ app-mirror/src/main/java/com/connect_screen/mirror/job/ExitAll.java
@@ -4,6 +4,8 @@ import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
+import android.os.Handler;
+import android.os.Looper;
 
 import com.connect_screen.mirror.BuildConfig;
 import com.connect_screen.mirror.State;
@@ -13,6 +15,9 @@ import com.connect_screen.mirror.shizuku.ShizukuUtils;
 
 public class ExitAll {
     public static void execute(Context context, boolean restart) {
+        if(true) {
+            return;
+        }
         if (SunshineService.instance != null) {
             SunshineService.instance.releaseWakeLock();
         }
@@ -27,7 +32,7 @@ public class ExitAll {
         }
         State.setMediaProjection(null);
         // 重启应用
-        if (restart) {
+        if (false) {
             PackageManager packageManager = context.getPackageManager();
             Intent intent = packageManager.getLaunchIntentForPackage(BuildConfig.APPLICATION_ID);
             ComponentName componentName = intent.getComponent();
@@ -35,7 +40,7 @@ public class ExitAll {
             // Required for API 34 and later
             // Ref: https://developer.android.com/about/versions/14/behavior-changes-14#safer-intents
             mainIntent.setPackage(context.getPackageName());
-            mainIntent.putExtra("DoNotAutoStartMoonlight", true);
+//            mainIntent.putExtra("DoNotAutoStartMoonlight", true);
             context.startActivity(mainIntent);
         }
 
@@ -59,12 +64,15 @@ public class ExitAll {
         if (context != null) {
             context.stopService(new Intent(context, SunshineService.class));
         }
-        // 退出应用进程
-        System.exit(0);
+        if (true) {
+            return ;
+        }
         try {
             android.os.Process.killProcess(android.os.Process.myPid());
         } catch(Throwable e) {
             // ignore
         }
+        // 退出应用进程
+        System.exit(0);
     }
 }
